requirements.txt
torch
torchvision
pillow
numpy
onnxruntime
opencv-python